api/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
python-docx
pypdf
google-generativeai
groq
httpx
pymongo[srv]